requirements.txt
datasets~=1.18.4
networkx~=2.6.3
numpy~=1.21.5
pandas~=1.3.5
pillow~=9.0.1
psutil~=5.9.1
pyarrow~=7.0.0
pytorch-lightning~=1.6.2
scikit-learn~=1.0.2
scipy~=1.7.3
seaborn~=0.11.2
tensorboard~=2.8.0
torchmetrics~=0.8.1
transformers~=4.17.0
wandb~=0.12.16
